refactor: LottoStore 를 LottoStoreService 로 변경
LottoStore를 도메인으로 두었으나, 로또를 판매하는 서비스에서 사용자가 돈을 주고 구매를 하는 기능이다. 실제로 멤버변수도 로또 자동생성기, 로또 수동생성기를 가지는데 이는 로또와 관련된 "도메인"이 아닌 "서비스"다.
