+++ src/main/java/lotto/Application.java
@@ -6,7 +6,7 @@ import lotto.domain.profit.LottoProfitCalculator;
 import lotto.domain.winning.LottoWinningSetPicker;
 import lotto.domain.winning.LottoWinningRankingCalculator;
 import lotto.domain.generator.LottoManualGenerator;
-import lotto.domain.store.LottoStore;
+import lotto.service.LottoStoreService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
@@ -16,12 +16,12 @@ public class Application {
         OutputView outputView = new OutputView();
         LottoAutoGenerator lottoAutoGenerator = new LottoAutoGenerator();
         LottoManualGenerator lottoManualGenerator = new LottoManualGenerator();
-        LottoStore lottoStore = new LottoStore(lottoAutoGenerator, lottoManualGenerator);
         LottoWinningSetPicker lottoWinningSetPicker = new LottoWinningSetPicker();
+        LottoStoreService lottoStoreService = new LottoStoreService(lottoAutoGenerator, lottoManualGenerator);
         LottoWinningRankingCalculator lottoWinningRankingCalculator = new LottoWinningRankingCalculator();
         LottoProfitCalculator lottoProfitCalculator = new LottoProfitCalculator();
 
-        LottoController lottoController = new LottoController(inputView, outputView, lottoStore, lottoWinningSetPicker, lottoWinningRankingCalculator, lottoProfitCalculator);
+        LottoController lottoController = new LottoController(inputView, outputView, lottoStoreService, lottoWinningSetPicker, lottoWinningRankingCalculator, lottoProfitCalculator);
         lottoController.run();
     }
 }
